+++ lib/parser.js
@@ -1,5 +1,6 @@
 var tokenizer = require('./tokenizer');
 var classHandlers = require('./handlers/classHandlers');
+var functionHandlers = require('./handlers/functionHandlers');
 
 var multiComment = false;
 var multiLine = false;
@@ -362,7 +363,7 @@ function handleWoof(parseContext)
   // module.exports = function x(a,b) {}
   if ( assignmentValue === 'such' )
   {
-    var functionStatement = handleSuch(parseContext);
+    var functionStatement = functionHandlers.handleSuch(parseContext);
     statement += functionStatement;
     // the closing wow will be in a new line
     return statement;
@@ -371,7 +372,7 @@ function handleWoof(parseContext)
   // module.exports = function (a,b) {}
   if ( assignmentValue === 'much')
   {
-    var anonStatement = handleLambda(parseContext);
+    var anonStatement = functionHandlers.handleLambda(parseContext);
     statement += anonStatement;
     // the closing wow will be in a new line
     return statement;
@@ -379,73 +380,6 @@ function handleWoof(parseContext)
 
   // TODO support other expressions
   throw new Error(`Invalid parse state! Expected a value but got: '${tokens[0]}' from chain: [${tokens}]. Allowed construct 'woof [<name> be] <value | <SUCH> | <MUCH> >'. ${parseInfo(parseContext)}`);
-}
-
-/**
- * Handles the such construct:
- *  such <function_name> [much]
- *
- * Produces:
- *  function <function_name> ([args]) {
- */
-function handleSuch(parseContext)
-{
-  expectToken('such', parseContext);
-
-  var tokens = parseContext.tokens;
-
-  // consume: such
-  tokens.shift();
-
-  var functionName = tokens.shift();
-  var statement = 'function ' + functionName;
-
-  // no args
-  if (!tokens[0])
-  {
-    return statement + ' () { \n';
-  }
-
-  // args have to be declared with much
-  if (tokens[0] !== 'much')
-  {
-    throw new Error(`Invalid parse state! Expected: 'much' but got: '${tokens[0]}' from chain: [${tokens}]. Allowed construct 'such <function_name> [much <args>]'. ${parseInfo(parseContext)}`);
-  }
-
-  statement += handleMuchArgs(parseContext);
-  return statement;
-}
-
-/**
- * Handles arguments to functions:
- *  much [args]
- *
- * Produces the argument part to a preceding function:
- *  ([args]) {
- *
- * Assuming that the function_name was encountered before this should ultimately produce:
- *  function([args]) {
- */
-function handleMuchArgs(parseContext)
-{
-  expectToken('much', parseContext);
-
-  // consume: much
-  parseContext.tokens.shift();
-
-  return declareArguments(parseContext);
-}
-
-/**
- * Handles lambda functions ( thanks @00Davo!):
- *  much [args]
- *
- * Produces:
- *  function (args) {
- */
-function handleLambda(parseContext)
-{
-  return 'function' + handleMuchArgs(parseContext);
 }
 
 /**
@@ -721,7 +655,7 @@ function handleWith(parseContext)
 
       // look ahead before consuming the tokens
       if (tokens[0] === 'much') { // lambda functions - thanks @00Davo!
-        var anonStatement = handleLambda(parseContext);
+        var anonStatement = functionHandlers.handleLambda(parseContext);
         statement += anonStatement;
         return statement;
       }
@@ -1057,7 +991,7 @@ function handleVery(parseContext) {
   }
 
   if (tokens[0] === 'much') {
-    var anonStatement = handleLambda(parseContext);
+    var anonStatement = functionHandlers.handleLambda(parseContext);
     statement += anonStatement;
     return statement;
   }
@@ -1205,7 +1139,7 @@ function parseStatement(parseContext) {
 
   // such function
   if (tokens[0] === 'such') {
-      return handleSuch(parseContext);
+      return functionHandlers.handleSuch(parseContext);
   }
 
   if ( tokens[0] === 'wow' || tokens[0] === 'wow&' )
